split commands to respective files
