+++ package.json
@@ -4,7 +4,7 @@
   "description": "CLI tool for parcel management.",
   "main": "./dist/index.js",
   "scripts": {
-    "start": "ts-node src/index.ts",
+    "start": "./dev.js",
     "build": "npm run clean && tsc && npm run linker:build",
     "clean": "rimraf dist && rimraf tmp",
     "linker:dev": "next",
@@ -43,6 +43,7 @@
     "decentraland-commons": "git+https://github.com/decentraland/commons.git#release",
     "decentraland-contracts": "git+https://github.com/decentraland/contracts.git#release",
     "fs-extra": "^5.0.0",
+    "get-installed-path": "^4.0.8",
     "ipfs-api": "^17.2.6",
     "koa": "^2.4.1",
     "koa-router": "^7.3.0",
