+++ src/utils/generate-html.ts
@@ -2,7 +2,7 @@ interface GeneratorSettings {
   withSampleScene?: boolean;
 }
 
-export default function generateHtml({ withSampleScene = false }: GeneratorSettings): string {
+export function generateHtml({ withSampleScene = false }: GeneratorSettings): string {
   const sampleScene = `<a-box position="-1 0.5 -3" rotation="0 45 0" color="#4CC3D9"></a-box>
       <a-sphere position="0 1.25 -5" radius="1.25" color="#EF2D5E"></a-sphere>
       <a-cylinder position="1 0.75 -3" radius="0.5" height="1.5" color="#FFC65D"></a-cylinder>
