+++ src/utils/is-dev.ts
@@ -1,8 +1,6 @@
 /**
  * Check if CLI is used in development mode.
  */
-const isDev =
+export const isDev:boolean =
   process.argv[1].indexOf('index') !== -1 ||
   process.argv[1].indexOf('dev') !== -1;
-
-export default isDev;
