+++ src/utils/scene-meta.ts
@@ -1,0 +1,30 @@
+const sceneMeta: DCL.SceneMetadata = {
+  display: {
+    title: 'My Land',
+    favicon: 'favicon_asset'
+  },
+  owner: '',
+  contact: {
+    name: '',
+    email: ''
+  },
+  main: 'scene',
+  tags: [],
+  scene: {
+    base: '',
+    parcels: []
+  },
+  communications: {
+    type: 'webrtc',
+    signalling: 'https://signalling-01.decentraland.org'
+  },
+  policy: {
+    contentRating: 'E',
+    fly: 'yes',
+    voiceEnabled: 'yes',
+    blacklist: [],
+    teleportPosition: ''
+  }
+};
+
+export default sceneMeta;
